requirements.txt
torch==2.9.1
torchvision==0.24.1
fastapi==0.121.1
uvicorn==0.38.0
starlette==0.49.3
python-multipart==0.0.6
pillow==12.0.0
